feat: allow "export VARIABLE=VALUE" syntax
Now the "parse()" functions return the set of exports along with
the environment map. This set is then used to export only specific
variables, if the global option "export" was not set (if it was set,
all variables are exported anyway).
The default variables are also used for marking variables as exports.
The example file is used to make sure that all necessary variables
have been exported.